Fix manifest cards * Make ManifestCard actually use the local static image correctly. * Separate ManifestCard into some smaller components. * Fix tests * Remove some debugging stuf that is no longer valid.
