+++ @ndlib/gatsby-theme-marble/src/components/Shared/Card/sx.js
@@ -28,12 +28,14 @@ module.exports = {
     }
   },
   imageBoarder: {
+    '& .gatsby-image-wrapper': {
+      width: '100vw',
+    },
     border: '1px solid',
     borderColor: 'gray.1',
     display: 'inline-block',
     margin: '0 auto',
     padding: '4px',
-    paddingBottom: '0',
   },
   figure: {
     margin: '0',
