+++ @ndlib/gatsby-theme-marble/src/components/Shared/Image/index.js
@@ -2,7 +2,6 @@
 // eslint-disable-next-line no-unused-vars
 import React from 'react'
 import PropTypes from 'prop-types'
-import Img from 'gatsby-image'
 import queryString from 'query-string'
 import noImage from 'assets/images/noImage.svg'
 import { jsx } from 'theme-ui'
@@ -19,29 +18,11 @@ const Image = ({
   alt, // alt text for the image
   title, // title attribute on iamge
 }) => {
-  let debug = false
-  try {
-    debug = queryString.parse(window.location.search).debug === 'true'
-  } catch {
-    debug = false
-  }
-  if (gatsbyImage) {
-    return (
-      <Img
-        fluid={gatsbyImage}
-        alt={alt || ''}
-        title={title}
-        fit='INSIDE'
-      />
-    )
-  }
-  // derive image from image service OR src OR use the default noImage
   const imageSrc = src || serviceURL(service, region, size) || noImage
-
   return (
-    <picture sx={debug ? sx.imageFail : null}>
+    <picture>
       <img
-        src={imageSrc}
+        src={gatsbyImage ? gatsbyImage.src : imageSrc}
         alt={alt || ''}
         title={title}
         sx={sx.fallBack}
