+++ @ndlib/gatsby-theme-marble/src/components/Shared/ManifestCard/index.js
@@ -1,12 +1,11 @@
 /** @jsx jsx */
-// eslint-disable-next-line no-unused-vars
-import React from 'react'
 import PropTypes from 'prop-types'
 import { useStaticQuery, graphql } from 'gatsby'
 import typy from 'typy'
 import { jsx } from 'theme-ui'
 import Card from 'components/Shared/Card'
 import TypeLabel from './TypeLabel'
+import ManifestCardChildren from './ManifestCardChildren'
 import sx from './sx'
 
 export const ManifestCard = (props) => {
@@ -36,8 +35,9 @@ export const ManifestCard = (props) => {
           name
           publicURL
           childImageSharp {
-            fluid(maxHeight: 250, maxWidth: 250, quality: 70) {
+            fluid(maxHeight: 250) {
               ...GatsbyImageSharpFluid
+              ...GatsbyImageSharpFluidLimitPresentationSize
             }
           }
         }
@@ -50,7 +50,6 @@ export const ManifestCard = (props) => {
     console.warn('Could not find manifest: ', props.iiifManifest)
     return null
   }
-  const children = figureOutChildren(props, item)
 
   const gatsbyImage = findGatsbyImage(item, allFile)
   let title = ''
@@ -68,58 +67,13 @@ export const ManifestCard = (props) => {
         imageService={typy(item, 'childrenMarbleIiifImage[0].service').safeString}
         {...props}
       >
-        {children}
+        <ManifestCardChildren
+          item={item}
+          parentProps={props}
+        />
       </Card>
       <TypeLabel type={item.display} />
     </div>
-  )
-}
-
-const findMetadata = (manifest, options) => {
-  if (!manifest.metadata) {
-    return []
-  }
-
-  return manifest.metadata.reduce((metaValue, row) => {
-    const label = typy(row, 'label').safeString.toLowerCase()
-
-    if (options.includes(label)) {
-      return metaValue.concat(row.value.join('<br/>'))
-    }
-
-    return metaValue
-  }, [])
-}
-
-export const figureOutChildren = (parentProps, item) => {
-  const dates = findMetadata(item, ['date', 'dates'])
-  let creator = ''
-  if (parentProps.highlight && parentProps.highlight.creator) {
-    creator = parentProps.highlight.creator[0]
-  } else {
-    creator = findMetadata(item, ['creator'])
-  }
-  return (
-    <React.Fragment>
-      {
-        parentProps.showCreator ? (
-          <p
-            sx={sx.lineStyle}
-            dangerouslySetInnerHTML={{ __html: creator }}
-          />
-        ) : null
-      }
-      {
-        parentProps.showDate ? (
-          <p
-            sx={sx.lineStyle}
-            dangerouslySetInnerHTML={{ __html: dates }}
-          />
-        ) : null
-      }
-      {parentProps.showSummary ? <div>{item.description}</div> : null}
-      {parentProps.children ? parentProps.children : null}
-    </React.Fragment>
   )
 }
 
@@ -130,14 +84,13 @@ const findItem = (manifestId, allMarbleItem) => {
 }
 
 const findGatsbyImage = (item, allFile) => {
-  // console.log(item.childrenMarbleIiifImage[0].name)
   if (!typy(item, 'childrenMarbleIiifImage[0].name').isString) {
     return null
   }
   const result = allFile.nodes.find(file => {
     return file.name.includes(typy(item, 'childrenMarbleIiifImage[0].name').safeString)
   })
-  return typy(result, 'childImageSharp.fixed').safeObject
+  return typy(result, 'childImageSharp.fluid').safeObject
 }
 
 ManifestCard.propTypes = {
