+++ @ndlib/gatsby-theme-marble/src/components/Shared/ManifestCard/test.js
@@ -1,8 +1,7 @@
 import React from 'react'
-import { shallow, mount } from 'enzyme'
+import { shallow } from 'enzyme'
 import { useStaticQuery } from 'gatsby'
-import { Styled } from 'theme-ui'
-import ManifestCard, { figureOutChildren } from './'
+import ManifestCard from './'
 import Card from 'components/Shared/Card'
 
 describe('ManifestCard', () => {
@@ -40,87 +39,5 @@ describe('ManifestCard', () => {
 
     const wrapper = shallow(<ManifestCard iiifManifest='badID' />)
     expect(wrapper.find(Card).exists()).toBeFalsy()
-  })
-
-  describe('figureOutChildren', () => {
-    const item = {
-      id: 'fancyId',
-      description: 'In 1814 we took a little trip,',
-
-      metadata: [
-        {
-          label: 'creator',
-          value: ['Johnny Horton', 'Andrew Jackson'],
-        },
-        {
-          label: 'date',
-          value: ['1814'],
-        },
-      ],
-    }
-    test('children no additional', () => {
-      const props = {
-        showCreator: false,
-        showDate: false,
-        children: <div className='child'>children are loud</div>,
-      }
-      const actual = figureOutChildren(props, item)
-      const wrapper = mount(actual)
-      expect(wrapper.find('.child').text()).toEqual('children are loud')
-    })
-
-    test('children + additional', () => {
-      const props = {
-        showCreator: true,
-        showDate: true,
-        showSummary: false,
-        children: <div className='child'>A song about an alligator.</div>,
-      }
-      const actual = figureOutChildren(props, item)
-      const wrapper = mount(actual)
-      expect(wrapper.find('p').at(0).html()).toContain('Johnny Horton<br>Andrew Jackson')
-      expect(wrapper.find('p').at(1).text()).toEqual('1814')
-      expect(wrapper.findWhere(c => {
-        return c.text() === 'In 1814 we took a little trip,'
-      }).exists()).toBeFalsy()
-      expect(wrapper.find('.child').text()).toEqual('A song about an alligator.')
-    })
-
-    test('no children', () => {
-      const props = {
-        showCreator: true,
-        showDate: true,
-        showSummary: true,
-      }
-      const actual = figureOutChildren(props, item)
-      const wrapper = mount(actual)
-      expect(wrapper.find('p').at(0).html()).toContain('Johnny Horton<br>Andrew Jackson')
-      expect(wrapper.find('p').at(1).text()).toEqual('1814')
-      expect(wrapper.findWhere(c => {
-        return c.text() === 'In 1814 we took a little trip,'
-      }).exists()).toBeTruthy()
-    })
-    describe('highlight search terms', () => {
-      test('highlight title', () => {
-        console.error = jest.fn()
-        useStaticQuery.mockImplementationOnce(() => {
-          return sq
-        })
-        const wrapper = shallow(<ManifestCard iiifManifest='someID' highlight={{ name: ['<em>label</em>'] }} />)
-        expect(wrapper.find(Card).html()).toContain('<em>label</em>')
-      })
-      test('creator highlight', () => {
-        const props = {
-          showCreator: true,
-          highlight: {
-            creator: ['<em>Author</em>'],
-          },
-        }
-
-        const actual = figureOutChildren(props, item)
-        const wrapper = mount(actual)
-        expect(wrapper.find('p').at(0).html()).toContain('<em>Author</em>')
-      })
-    })
   })
 })
